requirements.txt
Pillow
send2trash
